Update 1.VirtualFunctions.md
Parameters for the name of the file in the constructors of File, Folder, ImageFile etc. should not be passed by reference, as usually they are invoked with rvalues, not lvalues.
Functions std::vector<Folder*>& GetFolders() and std::vector<File*>& GetFiles() cannot be const, because we are returning member of the class by reference and thus we can change the member, which is not possible with const.